requirements.txt
Flask==2.2.3
numpy==1.22.4
openai==0.27.2
pandas==1.5.2
pymongo==4.3.3
python-dotenv==1.0.0
tiktoken==0.3.3
gunicorn
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
